Add files via upload
This is a single project activity of capacity 5 MW (i.e. AC capacity, which is equivalent to 6.5 MW of DC capacity) which is grid connected solar PV power generation project in the District Solapur, in the state of Maharashtra, in India. The purpose of this power is for captive consumption. The promoter of the project activity is ‘Vinati Organics Limited’ (herein after called as Project Proponent or PP). This project is an operational activity with GHG Reduction via renewable energy generation.